refactor: extract findFirstNodeOfKind util
