+++ lib/util/findFirstNodeOfKind.ts
@@ -1,0 +1,20 @@
+import ts from 'typescript';
+
+export const findFirstNodeOfKind = (root: ts.Node, kind: ts.SyntaxKind) => {
+  let result: ts.Node | undefined;
+  const visitor = (node: ts.Node) => {
+    if (result) {
+      return;
+    }
+
+    if (node.kind === kind) {
+      result = node;
+      return;
+    }
+    ts.forEachChild(node, visitor);
+  };
+
+  ts.forEachChild(root, visitor);
+
+  return result;
+};
